Allow XSLT preprocessing of MVA trainer description files. This is currently optional (and activated by parameter -x for mvaTreeTrainer or untracked.bool useXSLT in CMSSW).
The stylesheet is currently under data/MVATrainer.xsl and does not
transform anything at the moment.  In the future this could be used
to allo automatic expansion of "virtual" processors, emitting for
instance multiple processors at once and connecting them accordingly.

There is one feature you already get for free:
You can include file snippets (which have to be valid XML files,
so one would need to adjust the .xsl to make it more clever):

<xi:include href="snippet.xml"/>